backend/requirements-dev.txt
amqp==5.0.6
asgiref==3.3.4
attrs==21.2.0
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.2
django-celery-beat==2.2.0
django-extensions==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.12.0
django-seed==0.2.2
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-spectacular==0.15.1
Faker==8.1.2
inflection==0.5.1
jsonschema==3.2.0
kombu==5.0.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
PyJWT==2.1.0
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
runscript==0.2.10
setproctitle==1.2.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
vine==5.0.0
wcwidth==0.2.5
